+++ src/models/common.js
@@ -54,7 +54,7 @@ class Common {
         }
     };
 
-    static findAll(tenantId, type) {
+    static getAllMasters(tenantId, type) {
         let sql = "SELECT *, DATE_SUB(createdOn, INTERVAL 5 HOUR) AS adjusted_createdOn, DATE_SUB(updatedOn, INTERVAL 5 HOUR) AS adjusted_updatedOn FROM common_master";
         if (tenantId) {
             sql += ` WHERE tenantId = '${tenantId}'`;
@@ -66,34 +66,34 @@ class Common {
                 sql += ` WHERE type = '${type}'`;
             }
         }
+        return sql
+    }
+
+    static findAll(tenantId, type) {
+        let sql = this.getAllMasters(tenantId, type);
         sql += " ORDER BY name";
         return db.execute(sql)
     };
 
     static findActiveAll(tenantId, type) {
-        let sql = "SELECT *, DATE_SUB(createdOn, INTERVAL 5 HOUR) AS adjusted_createdOn, DATE_SUB(updatedOn, INTERVAL 5 HOUR) AS adjusted_updatedOn FROM common_master WHERE status = 1";
-        if (tenantId) {
-            sql += ` AND tenantId = '${tenantId}'`;
-        }
-        if (type) {
-            sql += ` AND type = '${type}'`;
-        }
+        let sql = this.getAllMasters(tenantId, type);
         sql += " ORDER BY name";
         return db.execute(sql);
     };
 
-    static findById(id) {
-        let sql = `SELECT * FROM common_master WHERE common_id = ${id}`;
+    static findById(tenantId, id) {
+        let sql = this.getAllMasters(tenantId);
+        sql += `AND common_id = ${id}`
         return db.execute(sql)
     };
 
-    static delete(commonId) {
-        let sql = `DELETE FROM common_master WHERE common_id = ${commonId}`;
+    static delete(tenantId, commonId) {
+        let sql = `DELETE FROM common_master WHERE tenantId = ${tenantId} AND common_id = ${commonId}`;
         return db.execute(sql)
     };
 
-    async update(id) {
-        let sql = `UPDATE common_master SET tenantId='${this.tenantId}',name='${this.name}',type='${this.type}',status='${this.status}',createdBy='${this.createdBy}',updatedBy='${this.updatedBy}',updatedOn='${this.dateandtime()}' WHERE common_id = ${id}`;
+    async update(tenantId, id) {
+        let sql = `UPDATE common_master SET name='${this.name}',type='${this.type}',status='${this.status}', updatedBy='${this.updatedBy}',updatedOn='${this.dateandtime()}' WHERE tenantId = ${tenantId} AND common_id = ${id}`;
         return db.execute(sql)
 
     };
